feat: namespace-scoped routing — all routes under /{handle}/
BREAKING CHANGE: All app routes now live under /{handle}/ prefix.
Old /workflows, /agents, /tasks etc. no longer exist.

Routing:
- All pages moved from (app)/ to (app)/[handle]/ via git mv
- /{handle}/workflows, /{handle}/agents, /{handle}/tasks, etc.
- /redirect client page resolves personal namespace → redirects
- Root / → /redirect, login/test-login → /redirect after auth

AppShell:
- Context switcher moved from bottom to top (under logo)
- Trigger shows active namespace name (org or "My profile")
- All nav links prefixed with /{handle} from URL
- Switching namespace navigates to /{ns.handle}/workflows

Internal links:
- All href/router.push updated across 22 pages + 10 components
- New useHandleFromPath() hook for shared components
- Server components use await params for handle
