+++ packages/platform-ui/src/app/(app)/[handle]/catalog/page.tsx
@@ -1,0 +1,6 @@
+import { redirect } from 'next/navigation';
+
+export default async function CatalogPage({ params }: { params: Promise<{ handle: string }> }) {
+  const { handle } = await params;
+  redirect(`/${handle}/agents`);
+}
