+++ packages/platform-ui/src/app/(app)/[handle]/monitoring/page.tsx
