+++ packages/platform-ui/src/app/(app)/[handle]/processes/page.tsx
@@ -1,0 +1,6 @@
+import { redirect } from 'next/navigation';
+
+export default async function ProcessesPage({ params }: { params: Promise<{ handle: string }> }) {
+  const { handle } = await params;
+  redirect(`/${handle}/workflows`);
+}
