--- packages/platform-ui/src/app/(app)/catalog/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function CatalogPage() {
-  redirect('/agents');
-}
