--- packages/platform-ui/src/app/(app)/processes/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function ProcessesPage() {
-  redirect('/workflows');
-}
