--- packages/platform-ui/src/app/(app)/workflows/[name]/runs/[runId]/report/page.tsx
@@ -1,70 +1,0 @@
-'use client';
-
-import { useParams } from 'next/navigation';
-import { useMemo } from 'react';
-import type { StepExecution, AuditEvent } from '@mediforce/platform-core';
-import { useProcessInstance, useSubcollection } from '@/hooks/use-process-instances';
-import { useAuditEvents } from '@/hooks/use-audit-events';
-import { useProcessDefinitionVersions } from '@/hooks/use-process-definitions';
-import { useWorkflowDefinitions } from '@/hooks/use-workflow-definitions';
-import { resolveDefinitionSteps } from '@/lib/resolve-definition-steps';
-import { RunReport } from '@/components/reports/run-report';
-
-type AuditEventWithId = AuditEvent & { id: string };
-
-export default function RunReportPage() {
-  const { name, runId } = useParams<{ name: string; runId: string }>();
-
-  const decodedName = name ? decodeURIComponent(name) : '';
-
-  const { data: instance, loading: instanceLoading } = useProcessInstance(runId ?? null);
-  const { data: stepExecutions } = useSubcollection<StepExecution>(
-    runId ? `processInstances/${runId}` : '',
-    'stepExecutions',
-  );
-  const { data: auditEvents } = useAuditEvents(runId ?? null);
-
-  const { versions: legacyVersions } = useProcessDefinitionVersions(decodedName);
-  const { definitions: workflowVersions } = useWorkflowDefinitions(decodedName);
-
-  const definitionSteps = useMemo(
-    () => resolveDefinitionSteps(instance, legacyVersions, workflowVersions),
-    [instance, legacyVersions, workflowVersions],
-  );
-
-  if (instanceLoading) {
-    return (
-      <div className="p-6 space-y-4 max-w-4xl mx-auto">
-        <div className="h-4 w-20 rounded bg-muted animate-pulse" />
-        <div className="h-8 w-1/2 rounded bg-muted animate-pulse" />
-        <div className="h-48 rounded bg-muted animate-pulse" />
-      </div>
-    );
-  }
-
-  if (!instance) {
-    return (
-      <div className="p-6 text-center text-sm text-muted-foreground">
-        Run not found.
-      </div>
-    );
-  }
-
-  if (instance.status !== 'completed') {
-    return (
-      <div className="p-6 text-center text-sm text-muted-foreground">
-        Report is only available for completed runs.
-      </div>
-    );
-  }
-
-  return (
-    <RunReport
-      instance={instance}
-      stepExecutions={stepExecutions}
-      auditEvents={auditEvents as AuditEventWithId[]}
-      definitionSteps={definitionSteps}
-      runDetailHref={`/workflows/${encodeURIComponent(decodedName)}/runs/${runId}`}
-    />
-  );
-}
