+++ packages/platform-ui/src/app/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation';
 
 export default function RootPage() {
-  redirect('/workflows');
+  redirect('/redirect');
 }
