+++ packages/platform-ui/src/components/agents/agent-run-list-table.tsx
@@ -7,6 +7,7 @@ import type { AgentRun } from '@mediforce/platform-core';
 import { ConfidenceBadge } from './confidence-badge';
 import { AutonomyBadge } from './autonomy-badge';
 import { cn } from '@/lib/utils';
+import { useHandleFromPath } from '@/hooks/use-handle-from-path';
 import type { LucideIcon } from 'lucide-react';
 
 function getPluginDisplay(pluginId: string): { Icon: LucideIcon; colorClass: string; label: string } {
@@ -57,6 +58,7 @@ export function AgentRunListTable({
   loading: boolean;
   processNameMap?: Map<string, string>;
 }) {
+  const handle = useHandleFromPath();
   return (
     <div className="rounded-md border overflow-auto">
       <table className="w-full text-sm">
@@ -97,7 +99,7 @@ export function AgentRunListTable({
                   <AutonomyBadge level={run.autonomyLevel} />
                 </td>
                 <td className="px-4 py-3">
-                  <Link href={`/workflows/${run.processInstanceId}`} className="text-xs text-muted-foreground hover:text-primary transition-colors">
+                  <Link href={`/${handle}/workflows/${run.processInstanceId}`} className="text-xs text-muted-foreground hover:text-primary transition-colors">
                     {processNameMap?.get(run.processInstanceId) ?? `${run.processInstanceId.slice(0, 8)}...`}
                   </Link>
                 </td>
