+++ packages/platform-ui/src/components/processes/runs-table.tsx
@@ -6,6 +6,7 @@ import { formatDistanceToNow } from 'date-fns';
 import type { ProcessInstance } from '@mediforce/platform-core';
 import { ProcessStatusBadge } from './process-status-badge';
 import { useUserDisplayNames } from '@/hooks/use-users';
+import { useHandleFromPath } from '@/hooks/use-handle-from-path';
 
 interface RunsTableProps {
   runs: ProcessInstance[];
@@ -17,18 +18,18 @@ interface RunsTableProps {
   emptyMessage?: string;
 }
 
-function defaultRunHref(run: ProcessInstance): string {
-  return `/workflows/${encodeURIComponent(run.definitionName)}/runs/${run.id}`;
-}
-
 export function RunsTable({
   runs,
   loading,
   showProcess = false,
-  runHref = defaultRunHref,
+  runHref,
   emptyMessage = 'No runs found.',
 }: RunsTableProps) {
+  const handle = useHandleFromPath();
   const userNames = useUserDisplayNames();
+  const effectiveRunHref = runHref ?? ((run: ProcessInstance) =>
+    `/${handle}/workflows/${encodeURIComponent(run.definitionName)}/runs/${run.id}`
+  );
   const headers = [
     ...(showProcess ? ['Workflow'] : []),
     'Run ID',
@@ -79,7 +80,7 @@ export function RunsTable({
               {showProcess && (
                 <td className="px-4 py-3 font-medium">
                   <Link
-                    href={`/workflows/${encodeURIComponent(run.definitionName)}`}
+                    href={`/${handle}/workflows/${encodeURIComponent(run.definitionName)}`}
                     className="hover:underline text-primary"
                   >
                     {run.definitionName}
@@ -114,7 +115,7 @@ export function RunsTable({
               </td>
               <td className="px-4 py-3">
                 <Link
-                  href={runHref(run)}
+                  href={effectiveRunHref(run)}
                   className="text-primary hover:underline inline-flex items-center gap-0.5 text-xs"
                 >
                   View <ChevronRight className="h-3 w-3" />
